fix(predicate): standardize lessThan on parseFloat to match greaterThan — Notidian-Notidian-a7k
filter.ts lessThan parsed operands with parseInt while greaterThan used
parseFloat, so decimals and radix-prefixed values were interpreted
inconsistently between the two operators. The isLessThanOrEqual /
isGreatThanOrEqual derivatives (defined as !greaterThan / !lessThan in
filterFnTypes) inherited that asymmetry.

Fix: change lessThan to parseFloat (single decided numeric-coercion
convention). Decimals (e.g. 1.5) are now honored on both sides, and
radix-prefixed values like '0x10' read as 0 under both operators
(parseFloat stops at 'x'), eliminating the parseInt-hex split.

NaN contract (pinned in-code + tests): a non-numeric operand parses to
NaN, and NaN < x / x < NaN is false, so a non-numeric value never
satisfies a numeric < or >; its <= / >= negation is therefore vacuously
true.

Regression tests in filter.test.ts: updated the lessThan block to the
parseFloat convention, added decimal/radix/trailing-unit cases, and a
new block exercising the isLessThanOrEqual / isGreatThanOrEqual
derivatives (boundary, decimals, radix, NaN) through filterFnTypes.
The Notidian-3fs follow-up note marks defect #2 fixed.

Gates: npm test --runInBand (107 suites / 2233 tests), tsc -noEmit
-skipLibCheck (exit 0), npm run build (exit 0).

+++ src/core/utils/contexts/predicate/filter.ts
@@ -75,7 +75,13 @@ export const lessThan: FilterFunction = (
   value: string,
   filterValue: string
 ): boolean => {
-  return parseInt(value) < parseInt(filterValue);
+  // Standardized on parseFloat to match greaterThan (single numeric-coercion
+  // convention) — decimals and radix prefixes are now interpreted identically
+  // by both operators, and by their isLessThanOrEqual/isGreatThanOrEqual
+  // derivatives (defined as !greaterThan/!lessThan in filterFnTypes). NaN
+  // contract: a non-numeric operand parses to NaN and NaN < x / x < NaN is
+  // false, so a non-numeric value never satisfies a numeric < (or >).
+  return parseFloat(value) < parseFloat(filterValue);
 };
 export const dateAfter: FilterFunction = (
   value: string,
